+++ RELEASE_NOTES.md
@@ -1,6 +1,13 @@
 # Release Notes
 
 
+## 1.2
+
+- `AlternateIconContext` is reneamed to `AlternateAppIconContext`.  
+- `AlternateIconListItem` is renamed to `AlternateAppIconListItem`.
+- `AlternateAppIconListItem` no longer has a size parameter.
+
+
 ## 1.1
 
 The library adds new depedendencies and a bunch of new functionality. 
@@ -20,15 +27,18 @@ The library adds new depedendencies and a bunch of new functionality.
 - `View` has a new `.standardLaunchOnboarding` modifier that performs a launch onboarding flow.
 
 
+
 ## 1.0.4
 
 - `AppInfo` has a new `appUrlScheme` property.
 - `AppUrls` has a new `app` property.
 
 
+
 ## 1.0.3
 
 - `AppInfo` now requires a `privacyUrl`.
+
 
 
 ## 1.0.2
@@ -40,11 +50,13 @@ The library adds new depedendencies and a bunch of new functionality.
 - `SocialMenuItem` view has a new initializer for icon customization
 
 
+
 ## 1.0.1
 
 - `AppInfo` is a new struct with app icons.
 - `AppUrls` is a new struct with app urls.
 - `SocialListItem` is a new view that can be used in the main menu.
+
 
 
 ## 1.0